refactor: fix comments
