+++ src/core/kilocode/task/message-utils.ts
